move process logs function to services util
